+++ src/ExposeErrorCode.php
@@ -5,6 +5,11 @@ namespace NineDigit\eKasa\Client;
 /**
  * Kódy chyby vrátený z Expose
  */
-class ExposeErrorCode {
+final class ExposeErrorCode
+{
     const TUNNEL_NOT_FOUND = 1;
+
+    private function __construct()
+    {
+    }
 }