+++ src/HttpMethod.php
@@ -8,4 +8,8 @@ final class HttpMethod
     public const POST = "POST";
     public const PUT = "PUT";
     public const DELETE = "DELETE";
+
+    private function __construct()
+    {
+    }
 }