+++ src/MediaTypeName.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace NineDigit\eKasa\Client;
+
+final class MediaTypeName
+{
+    public const APPLICATION_JSON = "application/json";
+    public const APPLICATION_PROBLEM_JSON = "application/problem+json";
+
+    private function __construct()
+    {
+    }
+}